requirements.txt
streamlit
praw
google-generativeai
python-dotenv